NEXT-37407 - Move migration to 6.5 namespace
