+++ src/Core/Migration/V6_5/Migration1715081559AdjustSentMailActionOnReviewSent.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Core\Migration\V6_6;
+namespace Shopware\Core\Migration\V6_5;
 
 use Doctrine\DBAL\ArrayParameterType;
 use Doctrine\DBAL\Connection;
